Remove unnecessary config files & update others (#15)
* No longer need .dir-locals.el & .vimrc

With the latest changes to `.editorconfig` (which were arrived at
through experiences in another project), the settings that were in
`.dir-locals.el` and `.vimrc` are no longer needed because they're
captured by `.editorconfig`.

* Add back max_line_length setting

After more discussion, the conclusion is that despite the issues with
`max_line_length`, it is handled as expected in the editors that many
people have been using.

* Adjust line length to 80 based on current project conventions