Removing unnecessary std::vector copy operations (#27)
When profiling Tesseract on experiments that include sampling several
shots, I noticed surprisingly significant time being spent in **copying
std::vector data structures**. The critical line, that consumed a
noticeable percentage of decoding time, was the line
`next_next_blocked_errs = next_blocked_errs` that is being called in
each iteration when generating a new search state from the current state
being processed. This line exists in order to create a local copy of the
vector for the next iteration and in order to ensure that changes on the
vector in the current iteration do not affect (are reversed) in the next
iteration.

The issue with this line is that it is necessary to perform copy if and
only if the `--at-most-two-errors-per-detector` flag is set. The logic
of the code that generates new search states and pushes them to the
priority queue is that the changes on the vector of blocked errors that
need to be reversed in next iteration happen only when this flag is
enabled. Performing copies of this vector when this flag is not enabled
is unnecessary and wastes a huge percentage of decoding time. Below are
graphs that show how much time was wasted for an example of each group
of protocol codes, when this flag was not enabled.

<br/>

![Screenshot 2025-06-03 8 52 22
AM](https://github.com/user-attachments/assets/eb0f712f-d123-4ebc-ac57-140c9c2caf60)

<br/>

![Screenshot 2025-06-03 8 53 10
AM](https://github.com/user-attachments/assets/57b08558-800b-4413-8eea-6ffa3ee2b152)

<br/>

![Screenshot 2025-06-03 8 53 56
AM](https://github.com/user-attachments/assets/45ac41c9-7354-4dec-aa40-333f89fbc1c3)

<br/>

![Screenshot 2025-06-03 8 54 48
AM](https://github.com/user-attachments/assets/ddbab7fd-26b5-4594-ad11-ca7d2c182891)

<br/>

![Screenshot 2025-06-03 8 56 20
AM](https://github.com/user-attachments/assets/de79a26c-c2e1-4bfd-ad49-f80e50d508fe)

<br/>

![Screenshot 2025-06-03 8 57 19
AM](https://github.com/user-attachments/assets/80e347d4-61c5-405c-8cb4-79ce9189068d)

---------

Signed-off-by: Dragana Grbic <draganaurosgrbic@gmail.com>
Co-authored-by: LaLeh <lalehbeni@google.com>