Move the bazel rules for the python wrapper to MODULE.bazel (#31)
For some reason using WORKSPACE makes it hard to link stim with the
python wrapper -or maybe I'm just writing the wrong bazel rules- This PR
moves the rules needed for the python wrapper to `MODULE.bazel` and adds
the rules needed to depend on pypi libraries.

note that bazel deprecated `WORKSPACE` and will drop support for it by
the end of the year https://bazel.build/external/migration but they
allow using both at the same time during migration. so this
PR starts the migration from `WORKSPACE` to `bzlmod`


part of https://github.com/quantumlib/tesseract-decoder/issues/17