link tesseract with stim and add tests for the python wrapper of common.h (#32)
some classes in stim are not exposed directly to python (e.g. DemTarget)
instead they are exposed through a child class (e.g. ExposedDemTarget)
... this forces us to do the conversion to be able to cross the
python/C++ boundary.

---

note that This PR doesn't actually link with stim instead the rule
"@stim_py//:stim" rebuilds the python wrapper for stim in a way that
allows us to use it. this is in order to unblock this project until I
figure out how to link to the `stim.so` file generated by
https://github.com/quantumlib/Stim/blob/f566b83c5da89ab94d7280178e2bd642350180c3/BUILD#L90

TODO: find the correct way to link to the `stim.so` file.

---

part of https://github.com/quantumlib/tesseract-decoder/issues/17