Keeps the `stim::DemInstructionType::DEM_LOGICAL_OBSERVABLE` in the dem during processing (#72)
This ensures that the number of logical observables from
`dem.count_observables()` remains correct, even if there
`DEM_LOGICAL_OBSERVABLE` observables not triggered by an error.

I also had it throw an exception for unrecognized instructions as we
really should be handling all the cases correctly now (and it would be
good for the user to know if we're not). Can switch back to cerr if
necessary though.