Update python documentation in readme (#91)
There isn't currently an example of how to use the decode_to_errors
function, even though it is a bit unintuitive, so I added this to the
readme.

Incidentally, I wonder if we should modify the decode_to_errors to
accept a bitstring of detector outcomes rather than indices of detection
events? That would make it consistent with the decode method.