Add workflow to label stale issues and PRs
As more and more issues and pull requests start to accumulate, it can be
helpful to automatically label stale issues and PRs and automatically
close them after a time.

This workflow labels issues with `stale` if the have not had any
activity for 90 days or more, and auto-closes them if they have not had
any further activity 60 days after that.