Use recursive search for pip dependencies
Tell Dependabot to look for pip dependency files in subdirectories.

Note: rather than limit it to `src/py` where `requirements_lock.txt` is
currently located, this will make it search in all directories so that
if the file locations change in the future, Dependabot is likely to be
able to still find the dependency files.